[CofenseTriage] Merge all test files into one
on-behalf-of: @Cofense <oss@cofense.com>
